+++ airo-camera-toolkit/airo_camera_toolkit/calibration/hand_eye_calibration.py
@@ -5,6 +5,8 @@ from typing import List, Optional
 
 import cv2
 from airo_camera_toolkit.calibration.collect_calibration_data import (
+    _start_teach_mode,
+    _stop_teach_mode,
     create_calibration_data_dir,
     save_calibration_sample,
 )
@@ -20,7 +22,6 @@ from airo_camera_toolkit.calibration.fiducial_markers import (
 from airo_camera_toolkit.interfaces import RGBCamera
 from airo_camera_toolkit.utils.image_converter import ImageConverter
 from airo_dataset_tools.data_parsers.camera_intrinsics import CameraIntrinsics
-from airo_robots.manipulators.hardware.ur_rtde import URrtde
 from airo_robots.manipulators.position_manipulator import PositionManipulator
 from airo_typing import HomogeneousMatrixType, OpenCVIntImageType
 from loguru import logger
@@ -64,11 +65,9 @@ def do_camera_robot_calibration(
     window_name = "Hand-eye calibration"
     cv2.namedWindow(window_name, cv2.WINDOW_NORMAL)
 
-    # For now, the robot is assumed to be a UR robot with RTDE interface, as we make use of the teach mode functions.
+    # For now, the robot is assumed to be a UR or RealMan robot, as we make use of the teach mode functions.
     # TODO: make this more generic by providing a teach mode function in the PositionManipulator interface?
-    if not isinstance(robot, URrtde):
-        raise NotImplementedError("Only UR robots are supported for now.")
-    robot.rtde_control.teachMode()
+    _start_teach_mode(robot)
 
     MIN_POSES = 3
     tcp_poses_in_base: List[HomogeneousMatrixType] = []
@@ -87,7 +86,7 @@ def do_camera_robot_calibration(
 
         key = cv2.waitKey(1)
         if key == ord("q"):
-            robot.rtde_control.endTeachMode()
+            _stop_teach_mode(robot)
             break
 
         if key == ord("s"):