+++ airo-camera-toolkit/airo_camera_toolkit/cli.py
@@ -16,14 +16,17 @@ def cli() -> None:
 @cli.command(name="hand-eye-calibration")
 @click.option("--mode", default="eye_in_hand", help="eye_in_hand or eye_to_hand")
 @click.option("--robot_ip", default="10.42.0.162", help="robot ip address")
+@click.option("--robot_type", default="ur", help="robot type: 'ur' or 'realman'")
 @click.option("--calibration_dir", type=click.Path(exists=False), help="directory to save the calibration data to.")
 @click_camera_options
-def calibrate_with_ur(
+def hand_eye_calibration(
     mode: str,
     robot_ip: str,
+    robot_type: str = "ur",
     calibration_dir: Optional[str] = None,
     camera_brand: Optional[str] = None,
     camera_serial_number: Optional[str] = None,
+
 ) -> None:
     """Do hand-eye calibration with a UR robot. Will open camera stream and visualize the detected board
     pose. Press S to capture pose, press Q to finish. Make sure the detections look good (corners/contours are
@@ -40,13 +43,19 @@ def calibrate_with_ur(
         * If residual error is low but the visualization looks wrong, you might have use the wrong mode (eye_in_hand
         or eye_to_hand). Try rerunning the calibration with the compute_calibration.py script.
     """
-    from airo_robots.manipulators.hardware.ur_rtde import URrtde
 
     aruco_dict = AIRO_DEFAULT_ARUCO_DICT
     charuco_board = AIRO_DEFAULT_CHARUCO_BOARD
-
-    robot = URrtde(robot_ip, URrtde.UR3_CONFIG)
-
+    robot=None
+    if robot_type == "ur":
+        from airo_robots.manipulators.hardware.ur_rtde import URrtde
+        robot = URrtde(robot_ip, URrtde.UR3_CONFIG)
+    elif robot_type == "realman":
+        from airo_robots.manipulators.hardware.realman import RealmanControl
+        robot = RealmanControl(robot_ip)
+    else:
+        raise NotImplementedError(f"Robot type {robot_type} is not supported yet. Only 'ur' and 'realman' are supported.")
+    
     camera = discover_camera(camera_brand, camera_serial_number)
     do_camera_robot_calibration(mode, aruco_dict, charuco_board, camera, robot, calibration_dir)